[FLINK-33422] Move Calc restore tests
